Issue #3891: reorganized doclet package inputs
